refactor(engine): tidy key-loading module layout
- Move the engine bind/destroy wiring and the exported dynamic-engine entry
  points out of lib.rs into their own engine_impl module, matching the
  one-file-per-module layout used by context/keyload/uri/logging.
- Move the integration-only masked-key generator into a dedicated
  integration module (it is a test helper gated behind the integration
  feature, not production engine API); update the masked-keygen consumer.
- Parse the URI key type via a FromStr impl on KeyType instead of a free
  parse_key_type function, with a doc comment on the conversion.
- Move the test-only helpers (write_key_material, new_test_engine) into the
  test module so their imports live at module scope rather than in the
  function bodies.
